pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "gemma-skill"
version = "0.4.0"
description = "Call Gemma 4 31B via Google AI Studio"
requires-python = ">=3.10"
dependencies = [
    "google-genai",
    "prompt-toolkit",
]

[project.scripts]
gemma = "gemma.cli:main"

[tool.setuptools]
packages = ["gemma"]
